+++ packages/dashboard/src/components/admin/user-list-card.tsx
@@ -24,7 +24,7 @@ import React from 'react';
 import { ConfirmationDialog, Loading, useAsync } from 'react-components';
 import { useNavigate } from 'react-router';
 
-import { AppControllerContext } from '../app-contexts';
+import { useAppController } from '../../hooks/use-app-controller';
 import { CreateUserDialog, CreateUserDialogProps } from './create-user-dialog';
 
 const ItemsPerPage = 20;
@@ -69,7 +69,7 @@ export function UserListCard({
   const [openDeleteDialog, setOpenDeleteDialog] = React.useState(false);
   const [deleting, setDeleting] = React.useState(false);
   const [openCreateDialog, setOpenCreateDialog] = React.useState(false);
-  const { showAlert } = React.useContext(AppControllerContext);
+  const { showAlert } = useAppController();
 
   const refresh = React.useCallback(async () => {
     if (!searchUsers) return;
@@ -143,7 +143,7 @@ export function UserListCard({
                       color="secondary"
                       startIcon={<DeleteIcon />}
                       className={classes.controlsButton}
-                      onClick={(ev) => {
+                      onClick={(ev: React.MouseEvent) => {
                         ev.stopPropagation();
                         setSelectedUser(u);
                         setOpenDeleteDialog(true);
@@ -164,7 +164,7 @@ export function UserListCard({
           page={page}
           rowsPerPage={ItemsPerPage}
           rowsPerPageOptions={[ItemsPerPage]}
-          onPageChange={(_, newPage) => setPage(newPage)}
+          onPageChange={(_: React.MouseEvent, newPage: number) => setPage(newPage)}
         />
       </TableContainer>
       {openDeleteDialog && (
