+++ packages/dashboard/src/components/index.ts
@@ -1,9 +1,4 @@
-export * from './admin';
-export * from './app-base';
-export * from './app-contexts';
-export * from './app-events';
-export * from './login-card';
 export * from '../micro-apps/map-app';
-export * from './private-route';
+export * from './app-events';
 export * from './rmf-dashboard';
 export * from './workspace';
