+++ packages/dashboard/src/hooks/use-settings.ts
@@ -1,0 +1,4 @@
+import { Settings } from '../services/settings';
+import { createDeferredContext } from './deferred-context';
+
+export const [useSettings, SettingsProvider] = createDeferredContext<Settings>();
